+++ crates/storage/db-api/src/lib.rs
@@ -78,8 +78,6 @@ pub use reth_storage_errors::db::{DatabaseError, DatabaseWriteOperation};
 pub mod models;
 mod scale;
 
-mod utils;
-
 pub use database::Database;
 
 mod unwind;
