+++ crates/storage/db-models/src/lib.rs
@@ -2,7 +2,7 @@
 
 /// Accounts
 pub mod accounts;
-pub use accounts::AccountBeforeTx;
+pub use accounts::{AccountBeforeTx, AddressStorageKey, BlockNumberAddress};
 
 /// Blocks
 pub mod blocks;
@@ -10,4 +10,12 @@ pub use blocks::{StoredBlockBodyIndices, StoredBlockWithdrawals};
 
 /// Client Version
 pub mod client_version;
+mod utils;
+
 pub use client_version::ClientVersion;
+
+pub mod sharded_key;
+pub use sharded_key::ShardedKey;
+
+pub mod storage_sharded_key;
+pub use storage_sharded_key::StorageShardedKey;
