+++ crates/storage/db/Cargo.toml
@@ -13,6 +13,7 @@ workspace = true
 
 [dependencies]
 # reth
+reth-codecs.workspace = true
 reth-db-api.workspace = true
 reth-primitives = { workspace = true, features = ["reth-codec"] }
 reth-primitives-traits.workspace = true
